digital interleaving

+++ src/nfc-lib/lib-lab/lab-tasks/src/main/cpp/tasks/LogicDeviceTask.cpp
@@ -226,7 +226,7 @@ struct LogicDeviceTask::Impl : LogicDeviceTask, AbstractTask
 
       // default parameters for DSLogic
       device->set(hw::LogicDevice::PARAM_OPERATION_MODE, hw::DSLogicDevice::OP_STREAM);
-      device->set(hw::LogicDevice::PARAM_LIMIT_SAMPLES, static_cast<unsigned int>(-1));
+      device->set(hw::LogicDevice::PARAM_LIMIT_SAMPLES, static_cast<unsigned long long>(-1));
 
       // setup sample rate
       if (config.contains("sampleRate"))
